refactor(event): rename components InputsRadio, FormRadios, FormInputImg to RadioGroup, FormRadioGroup, FormUploadFile
