+++ frontend/src/components/FormRadioGroup.tsx
@@ -1,36 +1,42 @@
 import React from 'react';
 import { classNames } from '../common/utils/utils';
-import InputsRadio, { InputsRadioProps } from './InputsRadio';
+import FormReadOnlyElement from './FormReadOnlyElement';
+import RadioGroup, { RadioGroupProps } from './RadioGroup';
 
-interface FormRadiosProps extends InputsRadioProps {
+interface FormRadiosProps extends RadioGroupProps {
   errorMessage?: string;
 }
 
-const FormRadios = ({
+const FormRadioGroup = ({
   errorMessage,
   readOnly,
   label,
+  value,
   options,
   className,
   helper,
   ...props
 }: FormRadiosProps) => {
   return readOnly ? (
-    <div>Read only to be added in edit</div>
+    <FormReadOnlyElement
+      label={label || ''}
+      value={value instanceof Array ? value.join(', ') : value}
+    />
   ) : (
-    <InputsRadio
+    <RadioGroup
+      {...props}
       options={options}
       label={label}
+      value={value}
       className={classNames(
         errorMessage
           ? 'border border-red-500 text-red-900 focus:ring-red-500 focus:border-red-500'
           : '',
         className || '',
       )}
-      {...props}
       helper={errorMessage ? <p className="text-red-500">{errorMessage}</p> : helper}
     />
   );
 };
 
-export default FormRadios;
+export default FormRadioGroup;
